backend/newspaper/requirements.txt
psutil === 5.7.2
selenium === 3.141.0
python-dateutil === 2.8.1
tldextract === 2.2.3
validators === 0.18.1